BRaVDA_v1.9
This update to BRaVDA includes a number of major changes.
The code has been rewritten to improve readability and has been slightly optimised.

BRaVDA will now automatically download MAS model runs and generate ensembles of solar wind speed for the generation of the prior state and prior error covariance matrix. As a result, premade MAS ensembles are no longer included as part of the BRaVDA package.

Observations from spacecraft are now averaged (as opposed to filtered by nearest neighbour) between timesteps to reduce the number of NaNs and data-gaps in the assimilation

Spacecraft positions are also better represented and account for deviations in radial position in addition to just longitude.

Potential latitudinal dependancy built into the original version of BRaVDA has been removed to save memory.

The initialisation of BRaVDA has changed, with BRaVDA now launched from the newly created startBRaVDA file, which invites the user to input variables that regularly need to be changed for DA experiments, such as output directory, start date and observations to assimilate.
A config file is added to store all the file locations/directories relative to the current directory and a file containing constants and values required by the underlying HUX solar wind model.
The original BRaVDA running script has now been packaged into a function and is run from the startBravda file.

Harriet Turner has contributed significantly to this update.